patched version 4.0.0
- added downstream service to be as close as possible to predecessor showcases
- upgraded to latest docker base images

+++ cnj-persistence-sql-backend-quarkus/src/main/java/group/msg/at/cloud/cloudtrain/core/boundary/TaskManagement.java
@@ -1,6 +1,7 @@
 package group.msg.at.cloud.cloudtrain.core.boundary;
 
 import group.msg.at.cloud.cloudtrain.adapter.persistence.jpa.repository.GenericRepository;
+import group.msg.at.cloud.cloudtrain.core.control.UserPermissionVerifier;
 import group.msg.at.cloud.cloudtrain.core.entity.Task;
 
 import javax.annotation.security.RolesAllowed;
@@ -21,28 +22,36 @@ import java.util.UUID;
 public class TaskManagement {
 
     @Inject
-    private GenericRepository repository;
+    GenericRepository repository;
+
+    @Inject
+    UserPermissionVerifier verifier;
 
     @NotNull
     public UUID addTask(@NotNull @Valid Task newTask) {
+        verifier.requirePermission("TASK_CREATE");
         newTask.setId(UUID.randomUUID());
         this.repository.addEntity(newTask);
         return newTask.getId();
     }
 
     public void modifyTask(@NotNull @Valid Task modifiedTask) {
+        verifier.requirePermission("TASK_UPDATE");
         this.repository.setEntity(modifiedTask);
     }
 
     public Task getTaskById(@NotNull UUID taskId) {
+        verifier.requirePermission("TASK_READ");
         return this.repository.getEntityById(Task.class, taskId);
     }
 
     public void removeTask(@NotNull UUID taskId) {
+        verifier.requirePermission("TASK_DELETE");
         this.repository.removeEntityById(Task.class, taskId);
     }
 
     public List<Task> getAllTasks() {
+        verifier.requirePermission("TASK_READ");
         return this.repository.queryEntities(Task.class, Task.QUERY_ALL, null);
     }
 }
